refactor: вынос функции разворота строки из компонента
