+++ src/types/elements.ts
@@ -1,0 +1,3 @@
+import {ElementStates} from "./element-states";
+
+export type TElement = { value: string, state: ElementStates };
